Match existing actors based on keymap
I abused the keymap so it holds, whenever a new actor is created on import:
- the repo id (if any) the actor is associated with
- the actor name
- the actor id
- the source name (KLUDGE ALERT: it reads /tmp/source-name to get source-name! use e.g.
                   `echo 'mysrcname' > /tmp/source-name` before imports)

Each time we encounter an actor name in the import, it will first look in the keymap
to see if there's already an existing actor with that name in the same repository
and source name. If not, create a new actor.

- I also modified getRepository so it works when the information object's parent isn't save yet.
- I also disabled nested set updates for info object save()s.

Various fixes for the CCA import

- Improve SQL query lookups so they escape quotes properly
- Trim newlines around names
- Disable <dao> and related digital object tag parsing for EAD

Remove annoying warnings